--- app/src/main/java/com/working_group/votingapp/ListActivity.java
@@ -1,14 +1,0 @@
-package com.working_group.votingapp;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class ListActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setTitle("一覧");
-        setContentView(R.layout.activity_list);
-    }
-}
