+++ app/src/main/java/com/working_group/votingapp/activity/ResultActivity.java
@@ -1,0 +1,25 @@
+package com.working_group.votingapp.activity;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
+
+import com.working_group.votingapp.R;
+
+public class ResultActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_result);
+        setTitle("結果画面");
+
+        Button returnButton = (Button) findViewById(R.id.return_listActivity);
+        returnButton.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                finish();
+            }
+        });
+    }
+}
