+++ app/src/main/res/layout/activity_list.xml
@@ -4,9 +4,19 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.working_group.votingapp.ListActivity">
+    tools:context="com.working_group.votingapp.activity.ListActivity">
 
     <ListView
         android:layout_width="368dp"
         android:layout_height="495dp" />
+
+    <Button
+        android:id="@+id/send_button"
+        android:layout_width="120dp"
+        android:layout_height="wrap_content"
+        android:layout_margin="10dp"
+        android:gravity="center"
+        android:onClick="resultAct"
+        android:text="@string/resultPage"
+        android:textSize="20sp" />
 </android.support.constraint.ConstraintLayout>
