+++ app/src/main/res/layout/activity_result.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.working_group.votingapp.activity.ResultActivity">
+
+    <Button
+        android:id="@+id/return_listActivity"
+        android:text="@string/return_listPage"
+        android:textSize="20sp"
+        android:layout_margin="10dp"
+        android:layout_width="120dp"
+        android:layout_height="wrap_content" />
+
+</android.support.constraint.ConstraintLayout>
